code/requirements.txt
# Core dependencies
chess==1.9.4
numpy==1.24.3
torch==2.0.1
python-dotenv==1.0.0
tqdm==4.65.0

# Visualization and UI
pygame==2.5.0
pygame-popup==0.9.1

# Utilities
setproctitle==1.3.2  # For naming processes during multiprocessing
matplotlib==3.7.1    # For plotting training metrics

# Optional - for performance monitoring and optimization
tensorboard==2.13.0  # For visualizing training progress
psutil==5.9.5        # For system monitoring (CPU/memory usage)

# Development tools
pytest==7.3.1        # For unit testing
black==23.3.0        # For code formatting
pylint==2.17.4       # For linting
pynvml==12.0.0       # For NVIDIA GPU monitoring (if applicable)